+++ crates/loro-internal/src/loro.rs
@@ -17,8 +17,8 @@ use crate::{
     change::Timestamp,
     container::{idx::ContainerIdx, IntoContainerId},
     encoding::{
-        decode_snapshot, encode_oplog, export_snapshot, export_snapshot_compressed,
-        export_snapshot_v0, parse_header_and_body, EncodeMode, ParsedHeaderAndBody,
+        decode_snapshot, encode_oplog, export_snapshot, export_snapshot_v0, parse_header_and_body,
+        EncodeMode, ParsedHeaderAndBody,
     },
     handler::TextHandler,
     handler::TreeHandler,
@@ -364,15 +364,6 @@ impl LoroDoc {
         ans
     }
 
-    pub fn export_from_compressed(&self, vv: &VersionVector) -> Vec<u8> {
-        self.commit_then_stop();
-        let oplog = self.oplog.lock().unwrap();
-        let ans = encode_oplog(&oplog, vv, EncodeMode::CompressedReorderedRle);
-        drop(oplog);
-        self.renew_txn_if_auto_commit();
-        ans
-    }
-
     #[inline(always)]
     pub fn export_from_v0(&self, vv: &VersionVector) -> Vec<u8> {
         self.commit_then_stop();
@@ -484,13 +475,6 @@ impl LoroDoc {
     pub fn export_snapshot(&self) -> Vec<u8> {
         self.commit_then_stop();
         let ans = export_snapshot(self);
-        self.renew_txn_if_auto_commit();
-        ans
-    }
-
-    pub fn export_snapshot_compressed(&self) -> Vec<u8> {
-        self.commit_then_stop();
-        let ans = export_snapshot_compressed(self);
         self.renew_txn_if_auto_commit();
         ans
     }