+++ crates/loro-wasm/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.7.2-alpha.4
+
+### Patch Changes
+
+- Fix encoding value err
+
 ## 0.7.2-alpha.3
 
 ### Patch Changes